Ver 5.2.3: creando el primo, falla al comprobar que es primo

+++ Pra05/THashCliente.h
@@ -45,7 +45,7 @@ private:
     
     
 public:
-    bool esprimo(unsigned long n);
+    bool esprimo(unsigned n);
     
     inline unsigned long djb2(unsigned char *str) {
     unsigned long hash = 5381;
